Update tangentsOfCircle.c
I will take a different approach here. Instead of creating a polygon of the shape I want, draw the arc, then close it with the 2 other lines, then make a bounding box around the shape, and if a pixel within that bounding box AND within the 4 different lines, then draw it.